添加删除实验室的后端接口

+++ sciman-main/src/main/java/com/sciman/dao/LabMapper.java
@@ -13,7 +13,7 @@ public interface LabMapper {
     List<Laboratory> getAllLabs();
 
     @Delete("delete from laboratory where id = #{id}")
-    Integer delete(Integer id);
+    Integer delete(Long id);
 
     List<LaboratoryView> queryLabIntros(
             String nameFilter
